libs/framework/eventservice/src/events/shoppingBag/addMultiple/index.test.ts
/* eslint-disable @typescript-eslint/unbound-method */
import {AddMultiple, AddMultipleCallback} from "."
import Emitter from "../../../eventEmitter"

jest.mock("../../../eventEmitter", () => ({
    emit: jest.fn(),
    listen: jest.fn(),
}))

describe("Given a Add_MultipleBag ESB", () => {
    describe("When publishing", () => {
        const mockPublishData = {
            data: "1",
            pageUrl: "2",
            deliveryWeeks: "3",
            isGrouped: true,
            groupType: "4",
        }

        beforeAll(() => {
            const getBackEsb = new AddMultiple()
            getBackEsb.publish(mockPublishData)
        })

        it("should call the common PublishData", () => {
            // eslint-disable-next-line no-unused-expressions
            expect(Emitter.emit).toHaveBeenCalledWith("SHOPPING_BAG_ADD_MULTIPLE", mockPublishData)
        })
    })

    describe("When subscribing", () => {
        const mockCallback = jest.fn()

        beforeAll(() => {
            const getBackEsb = new AddMultiple()
            getBackEsb.subscribe(mockCallback)
        })

        it("should call the common PublishData", () => {
            // eslint-disable-next-line no-unused-expressions
            expect(Emitter.listen).toHaveBeenCalledWith("SHOPPING_BAG_ADD_MULTIPLE", mockCallback)
        })
    })
})

describe("Given a AddMultipleCallback ESB", () => {
    describe("When publishing", () => {
        const mockPublishData = {
            success: true,
            data: {},
            status: 123,
            textStatus: "test",
        }

        beforeAll(() => {
            const getBagCallbackEsb = new AddMultipleCallback()
            getBagCallbackEsb.publish(mockPublishData)
        })

        it("should call the common PublishData", () => {
            // eslint-disable-next-line no-unused-expressions
            expect(Emitter.emit).toHaveBeenCalledWith("SHOPPING_BAG_ADD_MULTIPLE_CALLBACK", mockPublishData)
        })
    })

    describe("When subscribing", () => {
        const mockCallback = jest.fn()

        beforeAll(() => {
            const getBagCallbackEsb = new AddMultipleCallback()
            getBagCallbackEsb.subscribe(mockCallback)
        })

        it("should call the common PublishData", () => {
            // eslint-disable-next-line no-unused-expressions
            expect(Emitter.listen).toHaveBeenCalledWith("SHOPPING_BAG_ADD_MULTIPLE_CALLBACK", mockCallback)
        })
    })
})
